Allowed only one instance of about page
I made a static variable to tell whether or not the about page is up.
This limits the number of pages that are up. Of course, if the about
page crashes, the user would be stuck without functionality. I want to
try out a singleton class (or any kosher alternative) in the future, but
for now this works.